feat: add --customChain flag
